add support for monomorphic typing, split out Query into own module
